fn-request-to-api/requirements.txt
functions-framework==3.8.2
google-cloud-storage==3.0.0
google-cloud-bigquery==3.30.0
pyarrow==16.1.0
pandas==2.2.3